First Working Mixed-Model

+++ processor/processor.py
@@ -43,7 +43,7 @@ class Processor(IO):
         self.init_environment()
 
         if self.arg.phase == 'train':
-            if os.path.isdir(self.arg.work_dir + '/train'):
+            if os.path.isdir(self.arg.work_dir + '/train') and not self.arg.resume: ##ADD
                 print('log_dir: ', self.arg.work_dir, 'already exist')
                 shutil.rmtree(self.arg.work_dir + '/train')
                 shutil.rmtree(self.arg.work_dir + '/val')
@@ -271,5 +271,6 @@ class Processor(IO):
         parser.add_argument('--model_args', action=DictAction, default=dict(), help='the arguments of model')
         parser.add_argument('--weights', default=None, help='the weights for network initialization')
         parser.add_argument('--ignore_weights', type=str, default=[], nargs='+', help='the name of weights which will be ignored in the initialization')
-
+        parser.add_argument('--resume', type=str2bool, default=False)###ADD
+        
         return parser
